+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Pre Terraform shell hook.
 - Post Terraform shell hook.
+- Terraform derivations accept nix store paths as source.
 
 ## [0.7.0] - 2020-08-27